Consolidate C tests
